+++ include/reone/game/object/creature.h
@@ -122,7 +122,7 @@ public:
     void applyDamageEffect(
         int amount,
         uint32_t damager,
-        const std::array<int16_t, 15> &damageAmounts) override;
+        const std::array<int, 15> &damageAmounts) override;
 
     void giveXP(int amount);
     void setXP(int xp);
