+++ src/libs/game/object/creature.cpp
@@ -991,7 +991,7 @@ void Creature::damage(int amount, uint32_t damager) {
 void Creature::applyDamageEffect(
     int amount,
     uint32_t damager,
-    const std::array<int16_t, 15> &damageAmounts) {
+    const std::array<int, 15> &damageAmounts) {
 
     if (_dead) {
         return;
