+++ src/libs/game/object/door.cpp
@@ -305,7 +305,7 @@ void Door::damage(int amount, uint32_t damager) {
 void Door::applyDamageEffect(
     int amount,
     uint32_t damager,
-    const std::array<int16_t, 15> &damageAmounts) {
+    const std::array<int, 15> &damageAmounts) {
 
     damager = damager ? damager : script::kObjectInvalid;
     setLastDamager(damager);
