+++ src/libs/game/object/placeable.cpp
@@ -207,7 +207,7 @@ void Placeable::damage(int amount, uint32_t damager) {
 void Placeable::applyDamageEffect(
     int amount,
     uint32_t damager,
-    const std::array<int16_t, 15> &damageAmounts) {
+    const std::array<int, 15> &damageAmounts) {
 
     damager = damager ? damager : script::kObjectInvalid;
     setLastDamager(damager);
